Deleted more items that should have been ignored
Git will ignore all files on the git ignore if that file was not already in the repository when it was placed on the git ignore.
Hence if a .db file was already in the repository then .db got placed on the git ignore the one that was already in the repository would not be ignored until deleted from repository.