Fixed Responder Chain bugs on MAC
We were overriding responders for windows given to us.
We now place ours first in the chain, handle the information we want, and pass everything through to any responder that may have already been in the chain.
This allows GInput and GBuffered input to function together on the same application.