Added per-platform section for GFile file testing
One section of GFile that coolects the name of every file in a directory is currently testing for the three names in a certain order. However, depending on the operating system, the files are created or stored or retrieved in different orders. Therefore a #ifdef section was created for the different platforms.