Fixed GInput mouseDelta
Add flags in the GetMousDelta. If two flags have same value, it means the fuction was call twice in the shot time. The mosueDelta should be same, but function might give the different mouseDelta, here the function returns redundant_opration, but still get the new mosueDelta. If two flags have different values, it means the mouseDelta was actully changed, so return SUCCESS.