FIXED: Windows Memory Leaks
GLog: Added "delete logFile" before rv for both CreateGLog & CreateGLogCustom. Memory was only manage for success cases but not fail cases.

GOpenGLSurface: Added decrementCount() calls in Unit Test. Made refCount var atomic & assigned to 1. (fixed the refCount problem) Added decrementation in ~GOpenGL() for gWnd & Added Incrementation in SetGWindow() for gWnd.

GWindow: Added the DecrementCount() on the unopenedWindow obj, in the Unit Test

GDirectX11SurfaceTest: Moved CreateGWindow() in one of the SECTIONS of the TESTCASE. Since it wasn't in a SECTION in a TESTCASE, the function was called for each SECTION in that TESTCASE. Added a decrementation for the remaining dx11Surface & gWnd_DX references.