Linux hang solved, WindowListener works
Changed while loops conditional in the LinuxwinProc

Got rid of any locks in the linuxwinProc doing that with the while loop condintional change seemed to fix the listener problem and the hang problem.

Got rid of some "oops" code on mac and windows side...NEED to go to those platforms and re-test the LastEvent unitTests on them.