Mac Check
Fixed some false positives with LastEvent tests. 
Mac's LastEvent works for all cases except for init case & and any Maximize case.

Need to find functions that trigger the events. EX: our move function was calling setFrame() which does the job bc you can just change the position with it too as well as size but that function it trigger only the resize events we have. Changed it to SetFrameTopLeft().

As for the Maximize case I couldn't find a function that triggers the windowDidEnterFullscreen event. The window goes to fullscreen but triggers other events.