Removed OGL ES Unit Test
OpenGL ES isn't supported by every PC driver it seems so it should not be part of the standard test suite. Also updated the version file to contain the updated version info. A new build bug seems to be on my windows 10 machine the auto generated commit ID & branch name don't appear to be being read from the repo by wolfram. Not a huge issue but annoying to be sure.