Snyced events test with windowListener on Linux
GWindowTests Test case: Sending events to listeners was failing due to a race condition where the windowTestValue would not be updated before it was checked. This was fixed by syncing with the windowListener by adding a test pending count and only checking the value after their are no pending test left.