Implemented Reliable GListener Event ordering
Due to the way std::map iterates over objects we could not guarantee first in first out ordering for Events for registered listeners. This threw a nasty wrench in the mix when trying to correctly release and reallocate direct2d resources attached to the IGDirectX11Surface and caused a huge headache.  I was able to solve this issue on the back-end for all listeners but these changes will need to be ported to all platforms and unit tested.