GController now uses simpler input codes for its' events
Due to the increase in information provide by event data and the lack of branching in Mac and Linux for handle different types of controllers. The old input code only provided redundant information and made use of the interface more complex. This also adds a filter id to be used to only connect desired controllers based on the controllerID provided at creation.