Mac Mojave CMake build confirmed but GWindow errors discovered.
The latest build of CMake work fine on Mojave but 4 tests in GWindow are currently failing. It seems the window can no longer be correctly configured. Unfortunately my current objective C skill are a bit lacking and it will take more work to find out what must be adjusted to work in the new OS.