Major fixes to how UI is handled on Mac OS
Upgrading to Mojave revealed that how the UI communication on Mac OS was unsafe. All UI communication on the OS must be done from the main thread. Added a routine to all classes that do this. Ideally we could clean this up and place it in a Utility file. Last know issue on Mojave is a permissions issue with GAudio in the console output but all tests do seem to pass.