Fixed reference counting bug GAudio/Sound/Music on Windows
When a new sound or music is created in the audio system, both audio and sound/music increments its count. since these are not user-owned references, the audio system now calls cleanup if any music or sound drops to a reference count of 1. the audio system then properly handles the deletion of the sound or music.