Refactored Linux hpp and fixed test battery crash
Renamed TestBeep to testBeep, which was causing Linux test battery to fail as Linux is case sensitive
Refactored and cleaned up Linux and Windows hpps
found bug in the reference counting system on Linux and started finding a fix. currently audio counter gets to 0 in some cases, and Audio is not getting properly deleted.